# swagger.yaml
...
swagger: "2.0"
info:
  description: |
    BaaSapi API is an HTTP API served by BaaSapi. It is used by the BaaSapi UI and everything you can do with the UI can be done using the HTTP API.

    # Authentication

    Most of the API baask8ss require to be authenticated as well as some level of authorization to be used.
    BaaSapi API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:
    ```
    Bearer eyJ...1YEE
    ```

    # Security

    Each API baask8s has an associated access policy, it is documented in the description of each baask8s.

    Different access policies are available:
    * Public access
    * Authenticated access
    * Restricted access
    * Administrator access

    ### Public access

    No authentication is required to access the baask8ss with this access policy.

    ### Authenticated access

    Authentication is required to access the baask8ss with this access policy.

    ### Restricted access

    Authentication is required to access the baask8ss with this access policy.
    Extra-checks might be added to ensure access to the resource is granted. Returned data might also be filtered.

    ### Administrator access

    Authentication as well as an administrator role are required to access the baask8ss with this access policy.

  version: "1.20.2"
  title: "BaaSapi API"
  contact:
...
host: "baasapi.domain"
basePath: "/api"
...
schemes:
...
paths:
...
  /baask8ss:
    get:
      tags:
      - "baask8ss"
      summary: "List baask8ss"
      description: |
        List all baask8ss based on the current user authorizations. Will
        return all baask8ss if using an administrator account otherwise it will
        only return authorized baask8ss.
        **Access policy**: restricted
      operationId: "Baask8sList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Baask8sListResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "baask8ss"
      summary: "Create a new baask8s"
      description: |
        Create a new baask8s that will be used to manage a Docker environment.
        **Access policy**: administrator
      operationId: "Baask8sCreate"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "Name"
        in: "formData"
        type: "string"
        description: "Name that will be used to identify this baask8s (example: my-baask8s)"
        required: true
      - name: "Baask8sType"
        in: "formData"
        type: "integer"
        description: "Environment type. Value must be one of: 1 (Docker environment), 2 (Agent environment) or 3 (Azure environment)"
        required: true
      - name: "URL"
        in: "formData"
        type: "string"
        description: "URL or IP address of a Docker host (example: docker.mydomain.tld:2375).\
          \ Defaults to local if not specified (Linux: /var/run/docker.sock, Windows: //./pipe/docker_engine)"
      - name: "PublicURL"
        in: "formData"
        type: "string"
        description: "URL or IP address where exposed containers will be reachable.\
          \ Defaults to URL if not specified (example: docker.mydomain.tld:2375)"
      - name: "GroupID"
        in: "formData"
        type: "string"
        description: "Baask8s group identifier. If not specified will default to 1 (unassigned)."
      - name: "TLS"
        in: "formData"
        type: "string"
        description: "Require TLS to connect against this baask8s (example: true)"
      - name: "TLSSkipVerify"
        in: "formData"
        type: "string"
        description: "Skip server verification when using TLS (example: false)"
      - name: "TLSSkipClientVerify"
        in: "formData"
        type: "string"
        description: "Skip client verification when using TLS (example: false)"
      - name: "TLSCACertFile"
        in: "formData"
        type: "file"
        description: "TLS CA certificate file"
      - name: "TLSCertFile"
        in: "formData"
        type: "file"
        description: "TLS client certificate file"
      - name: "TLSKeyFile"
        in: "formData"
        type: "file"
        description: "TLS client key file"
      - name: "AzureApplicationID"
        in: "formData"
        type: "string"
        description: "Azure application ID. Required if baask8s type is set to 3"
      - name: "AzureTenantID"
        in: "formData"
        type: "string"
        description: "Azure tenant ID. Required if baask8s type is set to 3"
      - name: "AzureAuthenticationKey"
        in: "formData"
        type: "string"
        description: "Azure authentication key. Required if baask8s type is set to 3"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Baask8s"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Baask8s management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8s management is disabled"
  /baask8ss/{id}:
    get:
      tags:
      - "baask8ss"
      summary: "Inspect an baask8s"
      description: |
        Retrieve details abount an baask8s.
        **Access policy**: restricted
      operationId: "Baask8sInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Baask8s identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Baask8s"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Baask8s not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8s not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "baask8ss"
      summary: "Update an baask8s"
      description: |
        Update an baask8s.
        **Access policy**: administrator
      operationId: "Baask8sUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Baask8s identifier"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Baask8s details"
        required: true
        schema:
          $ref: "#/definitions/Baask8sUpdateRequest"
      responses:
        200:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        404:
          description: "Baask8s not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8s not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Baask8s management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8s management is disabled"
    delete:
      tags:
      - "baask8ss"
      summary: "Remove an baask8s"
      description: |
        Remove an baask8s.
        **Access policy**: administrator
      operationId: "Baask8sDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Baask8s identifier"
        required: true
        type: "integer"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Baask8s not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8s not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Baask8s management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8s management is disabled"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  TLSConfiguration:
    type: "object"
    properties:
      TLS:
        type: "boolean"
        example: true
        description: "Use TLS"
      TLSSkipVerify:
        type: "boolean"
        example: false
        description: "Skip the verification of the server TLS certificate"
      TLSCACertPath:
        type: "string"
        example: "/data/tls/ca.pem"
        description: "Path to the TLS CA certificate file"
      TLSCertPath:
        type: "string"
        example: "/data/tls/cert.pem"
        description: "Path to the TLS client certificate file"
      TLSKeyPath:
        type: "string"
        example: "/data/tls/key.pem"
        description: "Path to the TLS client key file"
  AzureCredentials:
    type: "object"
    properties:
      ApplicationID:
        type: "string"
        example: "eag7cdo9-o09l-9i83-9dO9-f0b23oe78db4"
        description: "Azure application ID"
      TenantID:
        type: "string"
        example: "34ddc78d-4fel-2358-8cc1-df84c8o839f5"
        description: "Azure tenant ID"
      AuthenticationKey:
        type: "string"
        example: "cOrXoK/1D35w8YQ8nH1/8ZGwzz45JIYD5jxHKXEQknk="
        description: "Azure authentication key"
...
  Baask8s:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Baask8s identifier"
      Name:
        type: "string"
        example: "my-baask8s"
        description: "Baask8s name"
      Type:
        type: "integer"
        example: 1
        description: "Baask8s environment type. 1 for a Docker environment, 2 for an agent on Docker environment or 3 for an Azure environment."
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of the Docker host associated to this baask8s"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable"
      GroupID:
        type: "integer"
        example: 1
        description: "Baask8s group identifier"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this baask8s"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this baask8s"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      TLSConfig:
        $ref: "#/definitions/TLSConfiguration"
      AzureCredentials:
        $ref: "#/definitions/AzureCredentials"
  Baask8sSubset:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Baask8s identifier"
      Name:
        type: "string"
        example: "my-baask8s"
        description: "Baask8s name"
      Type:
        type: "integer"
        example: 1
        description: "Baask8s environment type. 1 for a Docker environment, 2 for an agent on Docker environment, 3 for an Azure environment."
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of the Docker host associated to this baask8s"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable"
      GroupID:
        type: "integer"
        example: 1
        description: "Baask8s group identifier"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this baask8s"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this baask8s"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      TLSConfig:
        $ref: "#/definitions/TLSConfiguration"
  GenericError:
    type: "object"
    properties:
      err:
        type: "string"
        example: "Something bad happened"
        description: "Error message"
...
  Baask8sListResponse:
    type: "array"
    items:
      $ref: "#/definitions/Baask8sSubset"
...
  Baask8sUpdateRequest:
    type: "object"
    properties:
      Name:
        type: "string"
        example: "my-baask8s"
        description: "Name that will be used to identify this baask8s"
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of a Docker host"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable.\
          \ Defaults to URL if not specified"
      GroupID:
        type: "integer"
        example: "1"
        description: "Group identifier"
      TLS:
        type: "boolean"
        example: true
        description: "Require TLS to connect against this baask8s"
      TLSSkipVerify:
        type: "boolean"
        example: false
        description: "Skip server verification when using TLS"
      TLSSkipClientVerify:
        type: "boolean"
        example: false
        description: "Skip client verification when using TLS"
      ApplicationID:
        type: "string"
        example: "eag7cdo9-o09l-9i83-9dO9-f0b23oe78db4"
        description: "Azure application ID"
      TenantID:
        type: "string"
        example: "34ddc78d-4fel-2358-8cc1-df84c8o839f5"
        description: "Azure tenant ID"
      AuthenticationKey:
        type: "string"
        example: "cOrXoK/1D35w8YQ8nH1/8ZGwzz45JIYD5jxHKXEQknk="
        description: "Azure authentication key"
...